confirmar exclusão

+++ read.php
@@ -36,7 +36,7 @@
                     <td> {$row['categoria']} </td>
                     <td>
                         <a href='update.php?id={$row['id_notas']}'>Editar</a> |
-                        <a href='delete.php?id={$row['id_notas']}'>Excluir</a>
+                        <a href='delete.php?id={$row['id_notas']}' onclick='return confirm(\"Tem certeza que deseja excluir esta nota?\");'>Excluir</a>
                     </td>
                 </tr>
                 <tr>
